Move getting/setting current workspace to backend.py
